Near completion +Socket support +Faster -Needs more speed

+++ src/edgedetector/Tasker.java
@@ -4,9 +4,9 @@ import java.awt.Rectangle;
 
 public class Tasker implements Runnable
 {
-	public double[][] result=null;
-	public double[][] input=null;
-	public double[][] input2=null;
+	public float[][] result=null;
+	public float[][] input=null;
+	public float[][] input2=null;
 	private Assignment assignment;
 	private Rectangle region;
 	public Tasker(Assignment assignment, Rectangle region)
@@ -23,10 +23,10 @@ public class Tasker implements Runnable
 				result=Conv.conv(input, Conv.GAUSS, region);
 				break;
 			case XDERIV:
-				result=Conv.conv(input, Conv.SOBEL_X, region, 0.0);
+				result=Conv.conv(input, Conv.SOBEL_X, region, 0.0f);
 				break;
 			case YDERIV:
-				result=Conv.conv(input, Conv.SOBEL_Y, region, 0.0);
+				result=Conv.conv(input, Conv.SOBEL_Y, region, 0.0f);
 				break;
 			case MAGNITUDE:
 				result=Conv.magnitude(input, input2, region);
